Update Puzzle.py
It is unclear whether the problem prioritizes putting the puzzle pieces in the bottom-most, left-most location or using the smallest rotation. Thus, two programs are included in this solution, neither of which provides the correct answers for the sample inputs. The master branch uses the smallest rotations, while the patch puts pieces in the bottom-most, left-most location they'll fit in. 

There are multiple items printed for debugging purposes, but the final answer is on the last line.